Funpay-Telegram-Stars-Without-KYC/requirements.txt
pyrofork
tgcrypto
python-dotenv
requests
colorama
qrcode
